requirements.txt
django==3.2.7
django-environ==0.7.0
psycopg2==2.9.1
mysqlclient==2.0.3
Pillow==8.3.2
djangorestframework==3.12.4
black==21.10b0
mkdocs==1.2.3
pymdown-extensions==9.0
mkdocs-material==7.3.6
markdown-include==0.6.0
mkdocs-git-revision-date-localized-plugin==0.10.1
django-cors-headers==3.10.0
jwt==1.3.1
django-searchable-encrypted-fields==0.1.9
django-filter==21.1
